refactor(fieldmap): extract control state and logic into ControlContext
